requirements.txt
flask==3.0.3
flask-cors==4.0.0
PyMuPDF==1.24.5
pytesseract==0.3.13
Pillow==10.3.0
pandas==2.2.3
python-dotenv==1.0.1
# tesseract-ocr is already installed on Vercel's environment
# Dependencies for Vercel environment
certifi>=2023.7.22
charset-normalizer>=3.3.2
idna>=3.6
requests>=2.31.0
urllib3>=2.0.7
